eliminar nice

+++ src/components/AreasGallery.vue
@@ -35,6 +35,7 @@
                 show: this.$store.getters.getUserType,
                 teste: this.$store.state.logged,
                 searchTxt: "",
+                bookingAreas: [],
                 reset: {
                     areaName: "",
                 }
@@ -53,6 +54,10 @@
             this.$store.state.currentArea = ({
                 areaName: "",
             })
+            if (localStorage.getItem("areaBookings")) {
+                this.$store.state.areaBookings = JSON.parse(localStorage.getItem("areaBookings"))
+            }
+            this.areaBookings = this.$store.state.areaBookings
         },
         methods: {
             deleteArea(name) {
@@ -62,24 +67,39 @@
                     showCancelButton: true,
                 }).then((result) => {
                     if (result.value) {
+                        let checker = false
                         for (let i = 0; i <= this.areas.length; i++) {
-                            if (this.areas[i].name === name) {
-                                for (let j in this.users) {
-                                    if (this.users[j].userType === "cliente") {
-                                        this.users[j].notifications.push({
-                                            txt: 'O Espaço' + this.areas[i].name +
-                                                ' foi removido da galeria de Espaços!'
-                                        })
-                                        localStorage.setItem("users", JSON.stringify(this.users));
-                                    }
+                            for (let b in this.areaBookings) {
+                                if (name == this.areaBookings[b].areaName && this.areaBookings[b].state ==
+                                    "Aprovado") {
+                                    Swal.fire({
+                                        icon: 'error',
+                                        text: 'Não Pode eliminar um Espaço com Reservas!',
+                                    })
+                                    checker = true
                                 }
-                                this.areas.splice(i, 1)
-                                localStorage.setItem("areas", JSON.stringify(this.areas));
-                                Swal.fire({
-                                    icon: 'success',
-                                    text: 'Espaço eliminado!',
-                                })
+
                             }
+                            if (checker == false) {
+                                if (this.areas[i].name === name) {
+                                    for (let j in this.users) {
+                                        if (this.users[j].userType === "cliente") {
+                                            this.users[j].notifications.push({
+                                                txt: 'O Espaço' + this.areas[i].name +
+                                                    ' foi removido da galeria de Espaços!'
+                                            })
+                                            localStorage.setItem("users", JSON.stringify(this.users));
+                                        }
+                                    }
+                                    this.areas.splice(i, 1)
+                                    localStorage.setItem("areas", JSON.stringify(this.areas));
+                                    Swal.fire({
+                                        icon: 'success',
+                                        text: 'Espaço eliminado!',
+                                    })
+                                }
+                            }
+
                         }
                     }
                 })
